pyproject.toml
[tool.poetry]
name = "foundry_smb3"
version = "0.5.1"
description = "The future of SMB3"
readme = "README.md"
authors = ["TheJoeSmo <joesmo.joesmo12@gmail.com>"]
license = "GPL3"
packages = [
    { include = "foundry" }
]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
PySide6 = "^6.3.0"
attrs = "^21.2.0"
pydantic = "^1.8.2"
pretty-errors = "^1.2.24"
qt-material = "^2.8.13"
single-source = "^0.2.0"
autodoc-pydantic = "^1.6.1"
black = "^22.3.0"
pyupgrade = "^2.37.3"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pytest = "^6.2.5"
flake8 = "^3.9.2"
hypothesis = "^6.21.4"
pytest-qt = "^4.0.2"
Sphinx = "^4.2.0"
numpydoc = "^1.1.0"
pydata-sphinx-theme = "^0.7.1"
GitPython = "^3.1.24"
isort = "^5.9.3"
pyinstaller = "^4.5.1"

[tool.poetry.scripts]
foundry = "foundry.main:start"

[tool.isort]
profile = "black"
known_first_party = "foundry"

[tool.black]
line-length = 120

[tool.pytest]
qt_api = "PySide6"
ignore = "tests/game/gfx/objects/"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
